feat: set up stand-alone mythril container image build
The build supports arm64 and amd64 with non-emulated cross-compile for
the foreign architecture, so it can run in CI environments that don't
provide arm64 runners, such as GitHub Actions.

The GitHub Actions build uses Sigstore Cosign to sign builds, which
makes it possible to verify that a given image was created by the GitHub
Actions runner running the workflow in this repo, which makes it
possible to trust & verify the images.