fix: allow PRs without docker/* changes to merge
Currently PRs without changes to docker/* files can't be merged, because
we have a required status check for the "Test & Build" job, but this job
is only started when docker/* files change.

We have to explicitly run a "Test & Build" job in such a case, GitHub
doesn't automatically infer that the build wasn't required. To do this
we have a no-op workflow that only runs a PR has no docker/* changes.

See the GitHub docs on this scenario:
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/troubleshooting-required-status-checks#handling-skipped-but-required-checks